# perf: switch store from RwLock to Mutex 1.74x performance increase ~98.8% error rate decrease