@KolbyML 2025-02-25
> Yeah, great find! I'm still not completely sure I understand why this change caused the improvement, but those numbers look much better

Yeah it is a little confusing and intuitive